rework portfolio's language proficiencies API impl to use PATCH for all state changes

+++ app/src/main/java/fi/helsinki/opintoni/service/portfolio/LanguageProficiencyService.java
@@ -1,9 +1,11 @@
 package fi.helsinki.opintoni.service.portfolio;
 
 import fi.helsinki.opintoni.domain.portfolio.PortfolioLanguageProficiency;
+import fi.helsinki.opintoni.dto.portfolio.LanguageProficienciesChangeDescriptorDto;
 import fi.helsinki.opintoni.dto.portfolio.LanguageProficiencyDto;
 import fi.helsinki.opintoni.repository.portfolio.LanguageProficiencyRepository;
 import fi.helsinki.opintoni.repository.portfolio.PortfolioRepository;
+import fi.helsinki.opintoni.security.authorization.PermissionChecker;
 import fi.helsinki.opintoni.service.converter.portfolio.LanguageProficiencyConverter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,14 +21,17 @@ public class LanguageProficiencyService {
     private final PortfolioRepository portfolioRepository;
     private final LanguageProficiencyRepository languageProficiencyRepository;
     private final LanguageProficiencyConverter languageProficiencyConverter;
+    private final PermissionChecker permissionChecker;
 
     @Autowired
     public LanguageProficiencyService(PortfolioRepository portfolioRepository,
                                       LanguageProficiencyRepository languageProficiencyRepository,
-                                      LanguageProficiencyConverter languageProficiencyConverter) {
+                                      LanguageProficiencyConverter languageProficiencyConverter,
+                                      PermissionChecker permissionChecker) {
         this.portfolioRepository = portfolioRepository;
         this.languageProficiencyRepository = languageProficiencyRepository;
         this.languageProficiencyConverter = languageProficiencyConverter;
+        this.permissionChecker = permissionChecker;
     }
 
     public List<LanguageProficiencyDto> findByPortfolioId(Long id) {
@@ -35,28 +40,45 @@ public class LanguageProficiencyService {
             .collect(Collectors.toList());
     }
 
-    public LanguageProficiencyDto addLanguageProficiency(Long portfolioId,
-                                                         LanguageProficiencyDto languageProficiencyDto) {
+    public void updateLanguageProficiencies(LanguageProficienciesChangeDescriptorDto changeDescriptor,
+                                            Long portfolioId,
+                                            Long userId) {
+        if (changeDescriptor.deletedIds != null) {
+            changeDescriptor.deletedIds.forEach(id -> deleteLanguageProficiency(id, userId));
+        }
+
+        if(changeDescriptor.updatedLanguageProficiencies != null) {
+            changeDescriptor.updatedLanguageProficiencies.forEach((dto) -> updateLanguageProficiency(dto, userId));
+        }
+
+        if(changeDescriptor.newLanguageProficiencies != null) {
+            changeDescriptor.newLanguageProficiencies.forEach((dto) -> addLanguageProficiency(portfolioId, dto));
+        }
+    }
+
+    private void addLanguageProficiency(Long portfolioId,
+                                                          LanguageProficiencyDto languageProficiencyDto) {
         PortfolioLanguageProficiency portfolioLanguageProficiency = new PortfolioLanguageProficiency();
-        portfolioLanguageProficiency.languageCode = languageProficiencyDto.languageCode;
+        portfolioLanguageProficiency.languageCode = languageProficiencyDto.language;
         portfolioLanguageProficiency.proficiency = languageProficiencyDto.proficiency;
         portfolioLanguageProficiency.portfolio = portfolioRepository.findOne(portfolioId);
 
-        return languageProficiencyConverter.toDto(
-            languageProficiencyRepository.save(portfolioLanguageProficiency)
-        );
+        languageProficiencyRepository.save(portfolioLanguageProficiency);
     }
 
-    public LanguageProficiencyDto updateLanguageProficiency(Long languageProficiencyId,
-                                                            LanguageProficiencyDto languageProficiencyDto) {
+    private void updateLanguageProficiency(LanguageProficiencyDto languageProficiencyDto, Long userId) {
+        validateOwnership(userId, languageProficiencyDto.id);
         PortfolioLanguageProficiency languageProficiency =
-            languageProficiencyRepository.findOne(languageProficiencyId);
+            languageProficiencyRepository.findOne(languageProficiencyDto.id);
         languageProficiency.proficiency = languageProficiencyDto.proficiency;
-
-        return languageProficiencyDto;
     }
 
-    public void deleteLanguageProficiency(Long languageProficiencyId) {
+    private void deleteLanguageProficiency(Long languageProficiencyId, Long userId) {
+        validateOwnership(userId, languageProficiencyId);
         languageProficiencyRepository.delete(languageProficiencyId);
+    }
+
+    private void validateOwnership(Long userId, Long languageProficiencyId) {
+        permissionChecker.verifyPermission(userId, languageProficiencyId, PortfolioLanguageProficiency.class);
     }
 }
